+++ main.rb
@@ -2,9 +2,9 @@ require './app'
 
 @main = Main.new
 
-def menu
 
-  my_lambda = lamba {
+def menu
+  my_lambda = lambda {
     puts 'welcome to school library!'
     puts ' '
     puts 'please choose an option by entering a number:'
